requirements.txt
# Django e REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Banco de dados
mysql-connector-python==9.3.0

# SSH/SFTP
paramiko==3.4.0
cryptography==41.0.7

# YAML
PyYAML==6.0.2

# HTTP requests
requests==2.32.3

# Frontend
PySide6==6.6.1

# Configuração
python-decouple==3.8

# Logging e utilities
colorama==0.4.6
termcolor==3.0.0

